+++ .buildpacks
@@ -1,0 +1,2 @@
+https://github.com/heroku/heroku-buildpack-nodejs.git#40214da800
+https://github.com/heroku/heroku-buildpack-python.git#fe3f12c430
