+++ .gitignore
@@ -1,8 +1,10 @@
+*.pyc
+*.gz
+npm-debug.log
 app/static/css/contribute/contribute.css
 app/static/css/sandstone/sandstone-resp.css
 app/static/css/sandstone/sandstone-resp.min.css
 app/static/css/misc.css
-app/contribute.json
-app/node_modules/
+node_modules/
 
 .DS_Store
