+++ Gruntfile.js
@@ -4,8 +4,8 @@ module.exports = function(grunt) {
         less: {
             sandstone: {
                 files: {
-                    'static/css/sandstone/sandstone-resp.css' : 'static/css/sandstone/sandstone-resp.less',
-                    'static/css/contribute/contribute.css' : 'static/css/contribute/contribute.less'
+                    'app/static/css/sandstone/sandstone-resp.css' : 'app/static/css/sandstone/sandstone-resp.less',
+                    'app/static/css/contribute/contribute.css' : 'app/static/css/contribute/contribute.less'
                 }
             },
             sandstone_prod: {
@@ -13,18 +13,18 @@ module.exports = function(grunt) {
                     compress: true
                 },
                 files: {
-                    'static/css/sandstone/sandstone-resp.min.css' : 'static/css/sandstone/sandstone-resp.less'
+                    'app/static/css/sandstone/sandstone-resp.min.css' : 'app/static/css/sandstone/sandstone-resp.less'
                 }
             },
             misc: {
                 files: {
-                    'static/css/misc.css' : 'static/css/misc.less'
+                    'app/static/css/misc.css' : 'app/static/css/misc.less'
                 }
             },
 
         },
         jshint: {
-            files: ['grunt.js', 'js/*.js'],
+            files: ['Gruntfile.js', 'app/static/*.js'],
             options: {
                 bitwise: true,
                 camelcase: true,
@@ -35,7 +35,7 @@ module.exports = function(grunt) {
                 latedef: true,
                 newcap: true,
                 noarg: true,
-                quotmark: "single",
+                quotmark: 'single',
                 regexp: true,
                 undef: true,
                 unused: true,
@@ -46,25 +46,46 @@ module.exports = function(grunt) {
         },
         csslint: {
             base_theme: {
-                src: "css/*.css",
+                src: 'app/static/css/*.css',
                 rules: {
-                    "empty-rules": 2,
-                    "fallback-colors": 2,
-                    "font-sizes": 2,
-                    "important": 2,
-                    "outline-none": 2,
-                    "vendor-prefix": 2,
-                    "zero-units": 2
+                    'empty-rules': 2,
+                    'fallback-colors': 2,
+                    'font-sizes': 2,
+                    'important': 2,
+                    'outline-none': 2,
+                    'vendor-prefix': 2,
+                    'zero-units': 2
                 }
             }
         },
         watch: {
             scripts: {
-                files: ['static/css/**/*.less'],
+                files: ['app/static/css/**/*.less'],
                 tasks: ['less'],
                 options: {
                     spawn: false,
                     livereload: true
+                }
+            }
+        },
+        compress: {
+            main: {
+                options: {
+                    mode: 'gzip',
+                    level: 9
+                },
+                expand: true,
+                cwd: 'app/static/',
+                dest: 'app/static/',
+                src: [
+                  '**/*.css',
+                  '**/*.js',
+                  '**/*.eot',
+                  '**/*.svg',
+                  '**/*.ttf',
+                ],
+                ext: function(fileExt) {
+                    return fileExt + '.gz';
                 }
             }
         }
@@ -74,8 +95,9 @@ module.exports = function(grunt) {
     grunt.loadNpmTasks('grunt-contrib-less');
     grunt.loadNpmTasks('grunt-css');
     grunt.loadNpmTasks('grunt-contrib-watch');
+    grunt.loadNpmTasks('grunt-contrib-compress');
 
     grunt.registerTask('default', ['less', 'watch']);
     grunt.registerTask('lintify', ['jshint', 'csslint']);
-    grunt.registerTask('prep_prod', ['less']);
+    grunt.registerTask('prep_prod', ['less', 'compress']);
 };
