+++ Procfile
@@ -1,0 +1,1 @@
+web: gunicorn app.wsgi:app --bind 0.0.0.0:$PORT --log-file=-
