--- app/requirements.txt
@@ -1,13 +1,0 @@
-# Used by the app.py
-Flask==0.10.1
-Jinja2==2.7.3
-MarkupSafe==0.23
-Werkzeug==0.9.6
-itsdangerous==0.24
-python-memcached==1.53
-jsonschema==2.3.0
-requests==2.3.0
-six==1.8.0
-pyOpenSSL
-ndg-httpsclient
-pyasn1
