+++ app/root_files/contribute.json
@@ -1,0 +1,1 @@
+../../contribute.json
