+++ requirements.txt
@@ -1,0 +1,22 @@
+Flask==0.10.1
+Flask-Cache==0.13.1
+Flask-Heroku-Cacheify==1.3
+Jinja2==2.7.3
+MarkupSafe==0.23
+Werkzeug==0.9.6
+cffi==0.9.2
+cryptography==0.8.1
+enum34==1.0.4
+gunicorn==19.3.0
+itsdangerous==0.24
+jsonschema==2.3.0
+ndg-httpsclient==0.3.3
+pyOpenSSL==0.14
+pyasn1==0.1.7
+pycparser==2.10
+pylibmc==1.4.1
+python-memcached==1.53
+redis==2.10.3
+requests==2.3.0
+six==1.8.0
+whitenoise==1.0.6
